first major update
Massive revamp that includes error and exception handling for requests, adds configuration options for lookback, saas or on-premise... Handles availabilty data gathering differently. No real performance improvements but the output is nice and clean on screen and debug provides cleaner and more useful info. Exits are handled properly and iteration is skipped when invalid or empty objects are found.